Implement more of the new testArbitraryConstaints() now that I have the test cases from NewPlan/ModuleTests/ConstraintLibrary's testCLib.cc via a small hook (setting an environment variable) added to its runTestCases() that I'll commit as soon as that checkout passes 'make everything'.
Two helper functions, readSet() and readInterval(), that build domains
from data read from the "test cases" file, are still incomplete but
are structured and used in a way that permits the test(s) to pass as
they are now.

The only other missing functionality that I can think of at the moment
is support for Europa (NewPlan) label sets, but those can be mapped to
EnumeratedDomains if need be, perhaps via a new LabelDomain class that
derives from EnumeratedDomain to retain the label names.